+++ hardware/MQTT.cpp
@@ -8,6 +8,7 @@
 #include "../main/SQLHelper.h"
 #include "../json/json.h"
 #include "../notifications/NotificationHelper.h"
+#include <inttypes.h>
 
 #define RETRY_DELAY 30
 
@@ -134,7 +135,7 @@ void MQTT::on_message(const struct mosquitto_message *message)
 	Json::Reader jReader;
 	std::string szCommand = "udevice";
 	std::vector<std::vector<std::string> > result;
-	unsigned long long idx = 0;
+	uint64_t idx = 0;
 
 	bool ret = jReader.parse(qMessage, root);
 	if (!ret)
@@ -155,9 +156,9 @@ void MQTT::on_message(const struct mosquitto_message *message)
 		if (!root["idx"].isInt64())
 			goto mqttinvaliddata;
 
-		idx = (unsigned long long)root["idx"].asInt64();
+		idx = (uint64_t)root["idx"].asInt64();
 		//Get the raw device parameters
-		result = m_sql.safe_query("SELECT HardwareID, DeviceID, Unit, Type, SubType FROM DeviceStatus WHERE (ID==%llu)", idx);
+		result = m_sql.safe_query("SELECT HardwareID, DeviceID, Unit, Type, SubType FROM DeviceStatus WHERE (ID==%" PRIu64 ")", idx);
 		if (result.empty())
 		{
 			_log.Log(LOG_ERROR, "MQTT: unknown idx received!");
@@ -171,8 +172,8 @@ void MQTT::on_message(const struct mosquitto_message *message)
 		if (!root["idx"].isInt64())
 			goto mqttinvaliddata;
 
-		idx = (unsigned long long)root["idx"].asInt64();
-		result = m_sql.safe_query("SELECT Name FROM Scenes WHERE (ID==%llu)", idx);
+		idx = (uint64_t)root["idx"].asInt64();
+		result = m_sql.safe_query("SELECT Name FROM Scenes WHERE (ID==%" PRIu64 ")", idx);
 		if (result.empty())
 		{
 			_log.Log(LOG_ERROR, "MQTT: unknown idx received!");
@@ -186,8 +187,8 @@ void MQTT::on_message(const struct mosquitto_message *message)
 		if (!root["idx"].isInt64())
 			goto mqttinvaliddata;
 
-		idx = (unsigned long long)root["idx"].asInt64();
-		result = m_sql.safe_query("SELECT Name FROM UserVariables WHERE (ID==%llu)", idx);
+		idx = (uint64_t)root["idx"].asInt64();
+		result = m_sql.safe_query("SELECT Name FROM UserVariables WHERE (ID==%" PRIu64 ")", idx);
 		if (result.empty())
 		{
 			_log.Log(LOG_ERROR, "MQTT: unknown idx received!");
@@ -506,13 +507,13 @@ void MQTT::ProcessMySensorsMessage(const std::string &MySensorsMessage)
 	ParseLine();
 }
 
-void MQTT::SendDeviceInfo(const int m_HwdID, const unsigned long long DeviceRowIdx, const std::string &DeviceName, const unsigned char *pRXCommand)
+void MQTT::SendDeviceInfo(const int m_HwdID, const uint64_t DeviceRowIdx, const std::string &DeviceName, const unsigned char *pRXCommand)
 {
 	boost::lock_guard<boost::mutex> l(m_mqtt_mutex);
 	if (!m_IsConnected)
 		return;
 	std::vector<std::vector<std::string> > result;
-	result = m_sql.safe_query("SELECT DeviceID, Unit, Name, [Type], SubType, nValue, sValue, SwitchType, SignalLevel, BatteryLevel, Options FROM DeviceStatus WHERE (HardwareID==%d) AND (ID==%llu)", m_HwdID, DeviceRowIdx);
+	result = m_sql.safe_query("SELECT DeviceID, Unit, Name, [Type], SubType, nValue, sValue, SwitchType, SignalLevel, BatteryLevel, Options FROM DeviceStatus WHERE (HardwareID==%d) AND (ID==%" PRIu64 ")", m_HwdID, DeviceRowIdx);
 	if (result.size() > 0)
 	{
 		std::vector<std::string> sd = result[0];
@@ -573,7 +574,7 @@ void MQTT::SendDeviceInfo(const int m_HwdID, const unsigned long long DeviceRowI
 		}
 
 		if (m_publish_topics & PT_floor_room) {
-			result = m_sql.safe_query("SELECT F.Name, P.Name, M.DeviceRowID FROM Plans as P, Floorplans as F, DeviceToPlansMap as M WHERE P.FloorplanID=F.ID and M.PlanID=P.ID and M.DeviceRowID=='%llu'", DeviceRowIdx);
+			result = m_sql.safe_query("SELECT F.Name, P.Name, M.DeviceRowID FROM Plans as P, Floorplans as F, DeviceToPlansMap as M WHERE P.FloorplanID=F.ID and M.PlanID=P.ID and M.DeviceRowID=='%" PRIu64 "'", DeviceRowIdx);
 			for(size_t i=0 ; i<result.size(); i++)
 			{
 				std::vector<std::string> sd = result[i];
@@ -588,10 +589,10 @@ void MQTT::SendDeviceInfo(const int m_HwdID, const unsigned long long DeviceRowI
 	}
 }
 
-void MQTT::SendSceneInfo(const unsigned long long SceneIdx, const std::string &SceneName)
+void MQTT::SendSceneInfo(const uint64_t SceneIdx, const std::string &SceneName)
 {
 	std::vector<std::vector<std::string> > result, result2;
-	result = m_sql.safe_query("SELECT ID, Name, Activators, Favorite, nValue, SceneType, LastUpdate, Protected, OnAction, OffAction, Description FROM Scenes WHERE (ID==%llu) ORDER BY [Order]", SceneIdx);
+	result = m_sql.safe_query("SELECT ID, Name, Activators, Favorite, nValue, SceneType, LastUpdate, Protected, OnAction, OffAction, Description FROM Scenes WHERE (ID==%" PRIu64 ") ORDER BY [Order]", SceneIdx);
 	if (result.empty())
 		return;
 	std::vector<std::vector<std::string> >::const_iterator itt;
@@ -635,7 +636,7 @@ void MQTT::SendSceneInfo(const unsigned long long SceneIdx, const std::string &S
 	else
 		root["Status"] = "Mixed";
 	root["Timers"] = (m_sql.HasSceneTimers(sd[0]) == true) ? "true" : "false";
-	unsigned long long camIDX = m_mainworker.m_cameras.IsDevSceneInCamera(1, sd[0]);
+	uint64_t camIDX = m_mainworker.m_cameras.IsDevSceneInCamera(1, sd[0]);
 	//root["UsedByCamera"] = (camIDX != 0) ? true : false;
 	if (camIDX != 0) {
 		std::stringstream scidx;
